handle flow emission using callback flow
